Rename project and fix intendation to tabs
