+++ src/Castle.Windsor.Extensions.DependencyInjection/NetCoreServiceScope.cs
@@ -1,0 +1,49 @@
+// Copyright 2004-2020 Castle Project - http://www.castleproject.org/
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+namespace Castle.Windsor.Extensions.DependencyInjection
+{
+	using System;
+	
+	using Microsoft.Extensions.DependencyInjection;
+
+	internal class NetCoreServiceScope : IServiceScope
+	{
+		private readonly IDisposable scope;
+		private readonly IServiceProvider serviceProvider;
+
+		public NetCoreServiceScope(IDisposable windsorScope, IServiceProvider serviceProvider)
+		{
+			if(windsorScope == null)
+			{
+				throw new ArgumentNullException(nameof(scope));
+			}
+
+			if(serviceProvider == null)
+			{
+				throw new ArgumentNullException(nameof(serviceProvider));
+			}
+
+			this.scope = windsorScope;
+			this.serviceProvider = serviceProvider;
+		}
+
+		public IServiceProvider ServiceProvider => serviceProvider;
+
+		public void Dispose()
+		{
+			scope.Dispose();
+		}
+	}
+}
