+++ src/Castle.Windsor.Extensions.DependencyInjection/README.md
