--- src/Castle.Windsor.Extensions.MsDependencyInjection/LoggerDependencyResolver.cs
@@ -1,42 +1,0 @@
-// Copyright 2004-2020 Castle Project - http://www.castleproject.org/
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-namespace Castle.Windsor.Extensions.MsDependencyInjection
-{
-    using Castle.Core;
-    using Castle.MicroKernel;
-    using Castle.MicroKernel.Context;
-    
-    using Microsoft.Extensions.Logging;
-
-    public class LoggerDependencyResolver : ISubDependencyResolver
-    {
-        private readonly IKernel kernel;
-
-        public LoggerDependencyResolver(IKernel kernel)
-        {
-            this.kernel = kernel;
-        }
-        public bool CanResolve(CreationContext context, ISubDependencyResolver contextHandlerResolver, ComponentModel model, DependencyModel dependency)
-        {
-            return dependency.TargetType == typeof(ILogger);
-        }
-
-        public object Resolve(CreationContext context, ISubDependencyResolver contextHandlerResolver, ComponentModel model, DependencyModel dependency)
-        {
-            var factory = kernel.Resolve<ILoggerFactory>();
-            return factory.CreateLogger(RegistrationAdapter.OriginalComponentName(model.Name));
-        }
-    }
-}
