--- src/Castle.Windsor.Extensions.MsDependencyInjection/WindsorScopedServiceProvider.cs
@@ -1,93 +1,0 @@
-// Copyright 2004-2020 Castle Project - http://www.castleproject.org/
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-namespace Castle.Windsor.Extensions.MsDependencyInjection
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Reflection;
-    
-    using Castle.Windsor;
-    
-    using Microsoft.Extensions.DependencyInjection;
-
-    internal class WindsorScopedServiceProvider : IServiceProvider, ISupportRequiredService, IDisposable
-    {
-        private readonly NetCoreScope scope;
-        private bool disposing = false;
-
-        private readonly IWindsorContainer container;
-        
-        public WindsorScopedServiceProvider(IWindsorContainer container)
-        {
-            this.container = container;
-            this.scope = NetCoreScope.Current;
-        }
-
-        public object GetService(Type serviceType)
-        {
-            using(var fs = new NetCoreScope.ForcedScope(scope))
-            {
-                return ResolveInstanceOrNull(serviceType, true);    
-            }
-        }
-
-        public object GetRequiredService(Type serviceType)
-        {
-            using(var fs = new NetCoreScope.ForcedScope(scope))
-            {
-                return ResolveInstanceOrNull(serviceType, false);    
-            }
-        }
-
-        public void Dispose()
-        {
-            if(scope is NetCoreRootScope)
-            {
-                if(!disposing)
-                {
-                    disposing = true;
-                    var disposableScope = scope as IDisposable;
-                    if(disposableScope != null)
-                    {
-                        disposableScope.Dispose();
-                    }
-                    container.Dispose();
-                }
-                
-            }
-        }
-        private object ResolveInstanceOrNull(Type serviceType, bool isOptional)
-        {
-            if (container.Kernel.HasComponent(serviceType))
-            {
-                return container.Resolve(serviceType);
-            }
-
-            if (serviceType.GetTypeInfo().IsGenericType && serviceType.GetGenericTypeDefinition() == typeof(IEnumerable<>))
-            {
-                var allObjects = container.ResolveAll(serviceType.GenericTypeArguments[0]);
-                Array.Reverse(allObjects);
-                return allObjects;
-            }
-
-            if (isOptional)
-            {
-                return null;
-            }
-
-            return container.Resolve(serviceType);
-        }
-    }
-}
